+++ project-api-tests/build.gradle
@@ -32,7 +32,8 @@ dependencies {
     testImplementation "org.ballerinalang:jballerina-debugger-integration-test:${ballerinaLangVersion}"
     testImplementation "org.ballerinalang:ballerina-test-utils:${ballerinaLangVersion}"
     testImplementation "org.eclipse.lsp4j:org.eclipse.lsp4j.debug:${lsp4jDebugVersion}"
-
+    testImplementation "com.fasterxml.jackson.core:jackson-databind:2.12.3"
+    
     jBallerinaDistribution project(path: ":ballerina", configuration: "jBallerinaDistribution")
     ballerinaDistribution project(path: ":ballerina", configuration: "ballerinaDistribution")
     ballerinaLinuxDistribution project(path: ":ballerina", configuration: "ballerinaLinuxDistribution")
