+++ project-api-tests/src/test/java/org/ballerina/projectapi/CentralTest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerina.central;
+package org.ballerina.projectapi;
 
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
@@ -38,27 +38,28 @@ import java.util.LinkedList;
 import java.util.Map;
 import java.util.Objects;
 
-import static org.ballerina.central.CentralTestUtils.BALLERINA_DEV_CENTRAL;
-import static org.ballerina.central.CentralTestUtils.BALLERINA_HOME_DIR;
-import static org.ballerina.central.CentralTestUtils.BALLERINA_TOML;
-import static org.ballerina.central.CentralTestUtils.MAIN_BAL;
-import static org.ballerina.central.CentralTestUtils.createSettingToml;
-import static org.ballerina.central.CentralTestUtils.deleteFiles;
-import static org.ballerina.central.CentralTestUtils.getBalaPath;
-import static org.ballerina.central.CentralTestUtils.getEnvVariables;
-import static org.ballerina.central.CentralTestUtils.getExecutableJarPath;
-import static org.ballerina.central.CentralTestUtils.getGenerateBalaLog;
-import static org.ballerina.central.CentralTestUtils.getPushedToCentralLog;
-import static org.ballerina.central.CentralTestUtils.getString;
-import static org.ballerina.central.CentralTestUtils.randomPackageName;
-import static org.ballerina.central.CentralTestUtils.updateFileToken;
-import static org.ballerina.central.TestUtils.DISTRIBUTIONS_DIR;
-import static org.ballerina.central.TestUtils.MAVEN_VERSION;
-import static org.ballerina.central.TestUtils.executeBuildCommand;
-import static org.ballerina.central.TestUtils.executeCommand;
-import static org.ballerina.central.TestUtils.executePullCommand;
-import static org.ballerina.central.TestUtils.executePushCommand;
-import static org.ballerina.central.TestUtils.executeSearchCommand;
+import static org.ballerina.projectapi.CentralTestUtils.BALLERINA_DEV_CENTRAL;
+import static org.ballerina.projectapi.CentralTestUtils.BALLERINA_HOME_DIR;
+import static org.ballerina.projectapi.CentralTestUtils.BALLERINA_TOML;
+import static org.ballerina.projectapi.CentralTestUtils.MAIN_BAL;
+import static org.ballerina.projectapi.CentralTestUtils.createSettingToml;
+import static org.ballerina.projectapi.CentralTestUtils.deleteFiles;
+import static org.ballerina.projectapi.CentralTestUtils.getBalaPath;
+import static org.ballerina.projectapi.CentralTestUtils.getEnvVariables;
+import static org.ballerina.projectapi.CentralTestUtils.getExecutableJarPath;
+import static org.ballerina.projectapi.CentralTestUtils.getGenerateBalaLog;
+import static org.ballerina.projectapi.CentralTestUtils.getPushedToCentralLog;
+import static org.ballerina.projectapi.CentralTestUtils.getString;
+import static org.ballerina.projectapi.CentralTestUtils.randomPackageName;
+import static org.ballerina.projectapi.CentralTestUtils.updateFileToken;
+import static org.ballerina.projectapi.TestUtils.DISTRIBUTIONS_DIR;
+import static org.ballerina.projectapi.TestUtils.MAVEN_VERSION;
+import static org.ballerina.projectapi.TestUtils.OUTPUT_CONTAIN_ERRORS;
+import static org.ballerina.projectapi.TestUtils.executeBuildCommand;
+import static org.ballerina.projectapi.TestUtils.executeCommand;
+import static org.ballerina.projectapi.TestUtils.executePullCommand;
+import static org.ballerina.projectapi.TestUtils.executePushCommand;
+import static org.ballerina.projectapi.TestUtils.executeSearchCommand;
 
 /**
  * Tests related central packaging.
@@ -86,7 +87,6 @@ public class CentralTest {
     private static final String TEST_PREFIX = "test_";
     private static final String ANY_PLATFORM = "any";
     private static final String JAVA11_PLATFORM = "java11";
-    private static final String OUTPUT_CONTAIN_ERRORS = "build output contain errors:";
     private static final String OUTPUT_NOT_CONTAINS_EXP_MSG = "build output does not contain expected message:";
 
     @BeforeClass()
